Move certificate signature to secrets
