+++ pkg/kclient/secrets_test.go
@@ -3,7 +3,6 @@ package kclient
 import (
 	"testing"
 
-	"github.com/openshift/odo/pkg/kclient/generator"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -62,7 +61,7 @@ func TestCreateTLSSecret(t *testing.T) {
 				}
 				return true, &secret, nil
 			})
-			selfsignedcert, err := generator.GenerateSelfSignedCertificate(tt.host)
+			selfsignedcert, err := GenerateSelfSignedCertificate(tt.host)
 			if err != nil {
 				t.Errorf("fkclient.GenerateSelfSignedCertificate unexpected error %v", err)
 			}
@@ -79,6 +78,33 @@ func TestCreateTLSSecret(t *testing.T) {
 						t.Errorf("secret name does not match the expected name, expected: %s, got %s", tt.objectMeta.Name, createdTLSSceret.Name)
 					}
 				}
+			}
+
+		})
+	}
+}
+
+func TestGenerateSelfSignedCertificate(t *testing.T) {
+
+	tests := []struct {
+		name string
+		host string
+	}{
+		{
+			name: "test1",
+			host: "1.2.3.4.nip.io",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+
+			cert, err := GenerateSelfSignedCertificate(tt.host)
+			if err != nil {
+				t.Errorf("Unexpected error %v", err)
+			}
+			if cert.CertPem == nil || cert.KeyPem == nil || len(cert.CertPem) == 0 || len(cert.KeyPem) == 0 {
+				t.Errorf("Invalid certificate created")
 			}
 
 		})
