+++ pkg/url/url.go
@@ -221,7 +221,7 @@ func Create(client *occlient.Client, kClient *kclient.Client, parameters CreateP
 				_, err := kClient.KubeClient.CoreV1().Secrets(kClient.Namespace).Get(defaultTLSSecretName, metav1.GetOptions{})
 				// create tls secret if it does not exist
 				if kerrors.IsNotFound(err) {
-					selfsignedcert, err := generator.GenerateSelfSignedCertificate(parameters.host)
+					selfsignedcert, err := kclient.GenerateSelfSignedCertificate(parameters.host)
 					if err != nil {
 						return "", errors.Wrap(err, "unable to generate self-signed certificate for clutser: "+parameters.host)
 					}
@@ -250,7 +250,13 @@ func Create(client *occlient.Client, kClient *kclient.Client, parameters CreateP
 
 		}
 
-		ingressParam := generator.IngressParams{ServiceName: serviceName, IngressDomain: ingressDomain, PortNumber: intstr.FromInt(parameters.portNumber), TLSSecretName: parameters.secretName, Path: parameters.path}
+		ingressParam := generator.IngressParams{
+			ServiceName:   serviceName,
+			IngressDomain: ingressDomain,
+			PortNumber:    intstr.FromInt(parameters.portNumber),
+			TLSSecretName: parameters.secretName,
+			Path:          parameters.path,
+		}
 		ingressSpec := generator.GenerateIngressSpec(ingressParam)
 		objectMeta := generator.CreateObjectMeta(parameters.componentName, kClient.Namespace, labels, nil)
 		// to avoid error due to duplicate ingress name defined in different devfile components
